fix(core): address Copilot review on PR #60
- Extract SQLITE_RESULT_CODE_NAMES + extractSqliteCode into a
  dependency-free db/sqliteCodes.ts leaf module, shared by
  serializedAdapter.ts and types.ts (WikiTransactionError), removing
  the duplicated "keep in sync manually" table.
- Replace Object.create-based delegation with a Proxy that binds
  delegated methods to the original adapter instance, so class-based
  adapters relying on ES #private fields or instanceof/identity checks
  keep working through guardReentrancy/withSerializedTransactions.
- Clarify SQLiteAdapter.withTransactionAsync JSDoc: serialization is
  provided by WikiMemory's wrapper, not guaranteed by a bare adapter
  implementation.
